requirements.txt
numpy==1.26.4
keras
matplotlib==3.8.3
tensorflow
pandas==2.2.1
deeplake
pillow
scikit-learn
opencv-python
kaggle